+++ kernel/ipc/local_socket.h
@@ -19,7 +19,7 @@ struct sockaddr_un {
 int sock_local_bind(struct socket* sock, const struct sockaddr *addr, socklen_t addrlen);
 int sock_local_listen (struct socket* sock, int backlog);
 int	sock_local_accept (struct socket *sock, struct socket **newsock, struct sockaddr *addr);
-int sock_local_recvfrom(struct socket* sock, void* buf, size_t len, int flags, struct sockaddr* src_addr, socklen_t* addrlen);
+ssize_t sock_local_recvfrom(struct socket* sock, void* buf, size_t len, int flags, struct sockaddr* src_addr, socklen_t* addrlen);
 int sock_local_sendto(struct socket* sock, const void *msg, size_t len, int flags, const struct sockaddr *to, socklen_t tolen);
 
 int local_sock_create(struct socket* s, int type, int protocol);
