+++ kernel/ipc/socket.h
@@ -67,7 +67,7 @@ struct socket_prot_ops {
     int	(*connect) (struct socket* sock, struct sockaddr* saddr, int sockaddr_len);
     int (*bind) (struct socket* sock, const struct sockaddr *addr, socklen_t addrlen);
     int (*listen) (struct socket* sock, int backlog);
-    int (*recvfrom) (struct socket* sock, void* buf, size_t len, int flags, struct sockaddr* src_addr, socklen_t* addrlen);
+    ssize_t (*recvfrom) (struct socket* sock, void* buf, size_t len, int flags, struct sockaddr* src_addr, socklen_t* addrlen);
     int (*sendto) (struct socket* sock, const void *msg, size_t len, int flags, const struct sockaddr *to, socklen_t tolen);
     int (*setsockopt) (struct socket *sock, int level, int optname, const void *optval, unsigned int optlen);
     int	(*accept) (struct socket *sock, struct socket **newsock, struct sockaddr *addr);
