+++ kernel/net/ipv4/tcp.h
@@ -63,7 +63,7 @@ int	sock_tcp4_connect(struct socket* sock, struct sockaddr* saddr, int sockaddr_
 int	sock_tcp4_accept(struct socket *sock, struct socket **newsock, struct sockaddr *addr);
 int sock_tcp4_bind(struct socket* sock, const struct sockaddr *addr, socklen_t addrlen);
 int sock_tcp4_listen(struct socket* sock, int backlog);
-int sock_tcp4_recvfrom(struct socket* sock, void* buf, size_t len, int flags, struct sockaddr* src_addr, socklen_t* addrlen);
+ssize_t sock_tcp4_recvfrom(struct socket* sock, void* buf, size_t len, int flags, struct sockaddr* src_addr, socklen_t* addrlen);
 int sock_tcp4_sendto(struct socket* sock, const void *msg, size_t len, int flags, const struct sockaddr *to, socklen_t tolen);
 
 #endif
